Everything is looking pretty
I just need to connect some of the things to the back end and then the basics of the site are good to go.  There are also some validation things that I still need to take care of/password encryption and stuff like that.  We can update the github accordingly later.